+++ src/main/java/com/ERP/sales_management/Repository/InvoiceRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface InvoiceRepository extends JpaRepository<Invoice, Long> {
+public interface InvoiceRepository extends JpaRepository<Invoice, Integer> {
 
-    List<Invoice> findByOrderId(Long orderId);
+    List<Invoice> findByOrderId(Integer orderId);
 }
